Move DeriveMultiSigAddress to utils
